+++ Project_BanVeTauTaiGasLacHong/src/main/java/dao/impl/LoaiChoDAOImpl.java
@@ -1,39 +1,38 @@
-package dao;
+package dao.impl;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityTransaction;
-import jakarta.persistence.Persistence;
 import model.LoaiCho;
 
 import java.util.List;
 
-public class LoaiChoDAO {
+public class LoaiChoDAOImpl {
+    private EntityManager em;
+
+    public LoaiChoDAOImpl(EntityManager em) {
+        this.em = em;
+    }
+
     public List<LoaiCho> getAllList() {
-        EntityManager em = Persistence.createEntityManagerFactory("mariadb")
-                .createEntityManager();
         EntityTransaction tx = em.getTransaction();
         List<LoaiCho> list = null;
-        tx.begin();
         try {
-            list = em.createQuery("select lc from LoaiCho lc", LoaiCho.class).getResultList();
+            tx.begin();
+            list = em.createQuery("SELECT lc FROM LoaiCho lc", LoaiCho.class).getResultList();
             tx.commit();
         } catch (Exception e) {
             tx.rollback();
             System.err.println("Lỗi khi lấy danh sách LoaiCho");
+            e.printStackTrace();
         }
         return list;
     }
 
     public LoaiCho getById(String id) {
-        EntityManager em = Persistence.createEntityManagerFactory("mariadb")
-                .createEntityManager();
-        EntityTransaction tr = em.getTransaction();
         return em.find(LoaiCho.class, id);
     }
 
     public boolean save(LoaiCho t) {
-        EntityManager em = Persistence.createEntityManagerFactory("mariadb")
-                .createEntityManager();
         EntityTransaction tr = em.getTransaction();
         try {
             tr.begin();
@@ -48,8 +47,6 @@ public class LoaiChoDAO {
     }
 
     public boolean update(LoaiCho t) {
-        EntityManager em = Persistence.createEntityManagerFactory("mariadb")
-                .createEntityManager();
         EntityTransaction tr = em.getTransaction();
         try {
             tr.begin();
@@ -64,13 +61,13 @@ public class LoaiChoDAO {
     }
 
     public boolean delete(String id) {
-        EntityManager em = Persistence.createEntityManagerFactory("mariadb")
-                .createEntityManager();
         EntityTransaction tr = em.getTransaction();
         try {
             tr.begin();
             LoaiCho t = em.find(LoaiCho.class, id);
-            em.remove(t);
+            if (t != null) {
+                em.remove(t);
+            }
             tr.commit();
             return true;
         } catch (Exception e) {
