+++ Project_BanVeTauTaiGasLacHong/src/main/java/dao/impl/VeTauDAOImpl.java
@@ -1,59 +1,59 @@
-package dao;
+package dao.impl;
 
-import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityTransaction;
-import jakarta.persistence.Persistence;
-import lombok.AllArgsConstructor;
-import model.LoaiCho;
-import model.Tau;
+import model.VeTau;
 
 import java.util.List;
-@AllArgsConstructor
-public class TauDAO {
+
+public class VeTauDAOImpl {
     private EntityManager em;
 
-    public List<Tau> getAllListT() {
+    public VeTauDAOImpl(EntityManager em) {
+        this.em = em;
+    }
+
+    public List<VeTau> getAllList() {
         EntityTransaction tx = em.getTransaction();
-        List<Tau> list = null;
-        tx.begin();
+        List<VeTau> list = null;
         try {
-            list = em.createQuery("select t from Tau t", Tau.class).getResultList();
+            tx.begin();
+            list = em.createQuery("SELECT vt FROM VeTau vt", VeTau.class).getResultList();
             tx.commit();
         } catch (Exception e) {
             tx.rollback();
-            tx.rollback();
+            System.err.println("Lỗi khi lấy danh sách VeTau");
+            e.printStackTrace();
         }
         return list;
     }
 
-    public Tau getById(String id) {
-        EntityTransaction tr = em.getTransaction();
-        return em.find(Tau.class, id);
+    public VeTau getById(String id) {
+        return em.find(VeTau.class, id);
     }
 
-    public boolean save(Tau t) {
+    public boolean save(VeTau t) {
         EntityTransaction tr = em.getTransaction();
         try {
             tr.begin();
             em.persist(t);
             tr.commit();
             return true;
-        }catch(Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             tr.rollback();
         }
         return false;
     }
 
-    public boolean update(Tau t) {
+    public boolean update(VeTau t) {
         EntityTransaction tr = em.getTransaction();
         try {
             tr.begin();
             em.merge(t);
             tr.commit();
             return true;
-        }catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             tr.rollback();
         }
@@ -64,15 +64,16 @@ public class TauDAO {
         EntityTransaction tr = em.getTransaction();
         try {
             tr.begin();
-            Tau t = em.find(Tau.class,id);
-            em.remove(t);
+            VeTau t = em.find(VeTau.class, id);
+            if (t != null) {
+                em.remove(t);
+            }
             tr.commit();
             return true;
-        }catch (Exception e) {
+        } catch (Exception e) {
             e.printStackTrace();
             tr.rollback();
         }
         return false;
     }
-
 }
